moves some functions to libp2p utils
