+++ network/p2p/libp2pUtils.go
@@ -11,6 +11,8 @@ import (
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-libp2p-core/peer"
+	"github.com/libp2p/go-libp2p-core/protocol"
+	"github.com/multiformats/go-multiaddr"
 
 	"github.com/onflow/flow-go/model/flow"
 )
@@ -122,4 +124,76 @@ func networkingInfo(identity flow.Identity) (string, string, crypto.PubKey, erro
 	}
 
 	return ip, port, lkey, nil
+}
+
+// MultiAddressStr receives a node ip and port and returns
+// its corresponding Libp2p MultiAddressStr in string format
+// in current implementation IP part of the node address is
+// either an IP or a dns4.
+// https://docs.libp2p.io/concepts/addressing/
+func MultiAddressStr(ip, port string) string {
+	parsedIP := net.ParseIP(ip)
+	if parsedIP != nil {
+		// returns parsed ip version of the multi-address
+		return fmt.Sprintf("/ip4/%s/tcp/%s", ip, port)
+	}
+	// could not parse it as an IP address and returns the dns version of the
+	// multi-address
+	return fmt.Sprintf("/dns4/%s/tcp/%s", ip, port)
+}
+
+// IPPortFromMultiAddress returns the IP/hostname and the port for the given multi-addresses
+// associated with a libp2p host
+func IPPortFromMultiAddress(addrs ...multiaddr.Multiaddr) (string, string, error) {
+
+	var ipOrHostname, port string
+	var err error
+
+	for _, a := range addrs {
+		// try and get the dns4 hostname
+		ipOrHostname, err = a.ValueForProtocol(multiaddr.P_DNS4)
+		if err != nil {
+			// if dns4 hostname is not found, try and get the IP address
+			ipOrHostname, err = a.ValueForProtocol(multiaddr.P_IP4)
+			if err != nil {
+				continue // this may not be a TCP IP multiaddress
+			}
+		}
+
+		// if either IP address or hostname is found, look for the port number
+		port, err = a.ValueForProtocol(multiaddr.P_TCP)
+		if err != nil {
+			// an IPv4 or DNS4 based multiaddress should have a port number
+			return "", "", err
+		}
+
+		//there should only be one valid IPv4 address
+		return ipOrHostname, port, nil
+	}
+	return "", "", fmt.Errorf("ip address or hostname not found")
+}
+
+func generateProtocolID(rootBlockID string) protocol.ID {
+	return protocol.ID(FlowLibP2PProtocolIDPrefix + rootBlockID)
+}
+
+// PeerAddressInfo generates the libp2p peer.AddrInfo for an identity given its node address.
+func PeerAddressInfo(identity flow.Identity) (peer.AddrInfo, error) {
+	ip, port, key, err := networkingInfo(identity)
+	if err != nil {
+		return peer.AddrInfo{}, fmt.Errorf("could not get translate identity to networking info %s: %w", identity.NodeID.String(), err)
+	}
+
+	addr := MultiAddressStr(ip, port)
+	maddr, err := multiaddr.NewMultiaddr(addr)
+	if err != nil {
+		return peer.AddrInfo{}, err
+	}
+
+	id, err := peer.IDFromPublicKey(key)
+	if err != nil {
+		return peer.AddrInfo{}, fmt.Errorf("could not extract libp2p id from key:%w", err)
+	}
+	pInfo := peer.AddrInfo{ID: id, Addrs: []multiaddr.Multiaddr{maddr}}
+	return pInfo, err
 }
